Pulls notifications out into separate handler
